% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_code_to_template.R
\name{add_code_to_template}
\alias{add_code_to_template}
\title{Add text to a template Rmd file}
\usage{
add_code_to_template(template_path, code)
}
\arguments{
\item{template_path}{path to file}

\item{code}{the R code to add to the template}
}
\value{
if success, add the provided code chunk into the template
}
\description{
Add text to a template Rmd file
}
\examples{
# add this code when the input is file path
code <- "rio::import(system.file('extdata', 'covid_hosp_uk_20201024.xlsx',
                                 package = 'episoap'))"

}
\keyword{internal}
