% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_code_to_template.R
\name{add_text_to_template}
\alias{add_text_to_template}
\title{Add text to a template Rmd file}
\usage{
add_text_to_template(template_path, text = "")
}
\arguments{
\item{template_path}{path to the template file}

\item{text}{the text to add to the template file}
}
\value{
The input file will be updated with the new text if it is
successfully added
}
\description{
Add text to a template Rmd file
}
\examples{
template_path <- create_template()
text <- "## Import the data"
}
\keyword{internal}
