% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{calculate_cfr}
\alias{calculate_cfr}
\title{Calculate CFR from incidence or linelist data}
\usage{
calculate_cfr(data, epidist = NULL, epidist_params = NULL)
}
\arguments{
\item{data}{The input data in the form of a linelist or incidence.}

\item{epidist}{an \code{epidist} object that contains the distribution parameters}

\item{epidist_params}{A list with the parameters to be used to get the delay
distribution. Possible values are:
\describe{
type, values, distribution, interval, shape, scale, meanlog, sdlog
}}
}
\value{
A \code{<data.frame>} with two rows containing respectively the CFR
across all and confirmed cases.
}
\description{
Calculate CFR from incidence or linelist data
}
\examples{
data <- read.csv(system.file("extdata", "Marburg_EqGuinea_linelist.csv",
                             package = "episoap"))
incidence_data <- convert_to_incidence(
  data = data,
  date_var_name = "Onset_week",
  cases_status_var_name = "Status",
  death_outcome = "dead",
  diagnosis_status_var_name = "Type",
  diagnosis_outcome = "confirmed"
)
cfr <- calculate_cfr(
  data = incidence_data,
  epidist = NULL
)
}
