% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{calculate_cfr_from_counts}
\alias{calculate_cfr_from_counts}
\title{Calculate CFR from count data}
\usage{
calculate_cfr_from_counts(total_cases, total_deaths, death_in_confirmed)
}
\arguments{
\item{total_cases}{A \code{<numeric>} that represents the total number of
known cases}

\item{total_deaths}{A \code{<numeric>} that represents the total number of
known deaths}

\item{death_in_confirmed}{A \code{<numeric>} that represents the number of
death among the confirmed cases.}
}
\value{
A \code{<tibble>} with one or two rows containing respectively the
overall CFR and CFR in confirmed cases.
}
\description{
Calculate CFR from count data
}
\examples{
cfr <- calculate_cfr_from_counts(
  total_cases = 1150,
  total_deaths = 12,
  death_in_confirmed = 2
)

}
