% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{calculate_delay_cfr}
\alias{calculate_delay_cfr}
\title{Title}
\usage{
calculate_delay_cfr(
  data,
  onset_death,
  onset_date_variable = "date_onset",
  deaths_date_variable = "date_death",
  date_scale = "day",
  death_status = "Status",
  outcome = "dead"
)
}
\arguments{
\item{outcome}{}
}
\description{
Title
}
\examples{
delay_cfr <- calculate_delay_cfr(
  data                   = data,
  onset_death            = onset_death,
  onset_date_variable    = "date_onset",
  deaths_date_variable   = "date_death",
  date_scale             = "day",
  death_status           = "Status",
  outcome                = "dead"
)
}
