% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{calculate_overall_cfr}
\alias{calculate_overall_cfr}
\title{Calculate CFR with uncertainty}
\usage{
calculate_overall_cfr(
  data,
  infection_type = "direct_contact",
  cases_status = "Status",
  outcomes = c("dead", "recovered"),
  diagnosis_status = "Type",
  diagnosis_outcome = "confirmed"
)
}
\arguments{
\item{data}{a \verb{data frame} or \code{linelist} object}

\item{infection_type}{the type of the infection}

\item{cases_status}{the name of the column with the information about whether
a case was dead or recovered}

\item{outcomes}{a \code{character} vector with the 2 outcomes about the cases
status. default is c("dead", "recovered")}

\item{diagnosis_status}{the name of the column with the cases diagnosis
outcome.}

\item{diagnosis_outcome}{the value, in the column with the cases diagnosis
outcome, used to represent the confirmed cases.}
}
\value{
a \code{list} with elements of type \code{vector}. Each vector contains the
estimated CFR value and its confident interval.
\enumerate{
\item cfr_total: the CFR among the total cases
\item cfr_confirmed_only: the CFR among the confirmed cases only
}
}
\description{
Calculate CFR with uncertainty
}
\examples{
overall_cfr <- calculate_overall_cfr(
  data              = system.file("extdata", "Marburg_EqGuinea_linelist.csv",
                                  package = "readepi"),
  infection_type    = "direct_contact",
  cases_status      = "Status",
  outcomes          = c("dead", "recovered"),
  diagnosis_status  = "Type",
  diagnosis_outcome = "confirmed"
)
}
