% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_type.R
\name{data_type}
\alias{data_type}
\title{Get Data Type from YAML Configuration}
\usage{
data_type(path = file.path(tempdir(), "config.yaml"))
}
\arguments{
\item{path}{Path to config file (default: tempdir()/config.yaml)}
}
\value{
Character string indicating data type
}
\description{
Get Data Type from YAML Configuration
}
