% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmissibility.R
\name{estimate_growth_rate}
\alias{estimate_growth_rate}
\title{Estimate the growth rate using {i2extra}}
\usage{
estimate_growth_rate(dat, group_var = NULL, alpha = 0.95)
}
\arguments{
\item{dat}{A data frame with the daily incidence data}

\item{group_var}{A string with the name of the column used to group the data.
This is \code{NULL} when the data is not grouped}

\item{alpha}{A numeric that represents the confidence interval level. Default
is 95\% (0.95)}
}
\value{
A data frame with the estimated growth rate
}
\description{
Estimate the growth rate using {i2extra}
}
\keyword{internal}
