% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{extract_params}
\alias{extract_params}
\title{Extract the distribution parameters from percentiles, median and range}
\usage{
extract_params(data, arg_list)
}
\arguments{
\item{data}{the input data frame or linelist}

\item{arg_list}{a list with the parameters}
}
\description{
Extract the distribution parameters from percentiles, median and range
}
\keyword{internal}
