% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{get_delay_distro_params}
\alias{get_delay_distro_params}
\title{Get the delay distribution parameters}
\usage{
get_delay_distro_params(
  tmp_data,
  type,
  values,
  distribution,
  shape,
  scale,
  meanlog,
  sdlog
)
}
\value{
an object an type \code{list} with the delay distribution parameters
}
\description{
Get the delay distribution parameters
}
\keyword{internal}
