% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{get_onset_to_death_distro}
\alias{get_onset_to_death_distro}
\title{Title}
\usage{
get_onset_to_death_distro(
  data,
  death_status = "Status",
  disease = "Marburg Virus Disease",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
onset_death <- get_onset_to_death_distro(
  data         = data,
  death_status = "Status",
  disease      = "Marburg Virus Disease",
  type         = "range",
  values       = c(8, 2, 16),
  distribution = "gamma"
)
}
