% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{get_params}
\alias{get_params}
\title{Get the distribution parameters of interest}
\usage{
get_params(arg_list)
}
\arguments{
\item{arg_list}{a list with the parameters}
}
\description{
Get the distribution parameters of interest
}
\keyword{internal}
