% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{get_sequential_dates}
\alias{get_sequential_dates}
\title{Convert date in input data for CFR calculation into sequential dates}
\usage{
get_sequential_dates(data)
}
\arguments{
\item{data}{a data frame with at least 2 columns: the \code{cases} and the \code{date}
they were recorded.}
}
\value{
a data frame with 3 columns named as 'date', 'cases', 'deaths'. This
is the format required by the \code{cfr::cfr_static()} function
}
\description{
Convert date in input data for CFR calculation into sequential dates
}
\keyword{internal}
