% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{get_severity}
\alias{get_severity}
\title{Estimate severity based on the user-provided arguments}
\usage{
get_severity(
  disease_name = NULL,
  data = NULL,
  account_for_delay = FALSE,
  epidist = NULL,
  ...
)
}
\arguments{
\item{disease_name}{a \code{character} with the name of the disease}

\item{data}{the input data}

\item{account_for_delay}{a \code{logical} used to specify whether to account for
the delay distribution or not.}

\item{epidist}{an object of class \code{epidist} with the delay distribution
parameters}

\item{...}{the extra parameters needed for some specific cases}
}
\value{
a list of data frames that contains the severity estimates.
}
\description{
Estimate severity based on the user-provided arguments
}
\examples{
# example code


}
