% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{get_severity_params}
\alias{get_severity_params}
\title{Extract the severity parameters from the input arguments}
\usage{
get_severity_params(parameters, severity_params)
}
\arguments{
\item{parameters}{a list of parameters to be populated with the severity
parameters}

\item{severity_params}{a list with the user-specified arguments to be used
for the CFR calculation}
}
\value{
the input list of parameters with extra arguments from the
'severity_params' object
}
\description{
Extract the severity parameters from the input arguments
}
\keyword{internal}
