% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmissibility.R
\name{get_si_distribution}
\alias{get_si_distribution}
\title{Get the serial interval distribution}
\usage{
get_si_distribution(
  disease_name,
  use_epiparameter_database = TRUE,
  si_mean = NULL,
  si_sd = NULL,
  si_dist = NULL
)
}
\arguments{
\item{disease_name}{A string with the disease name.}

\item{use_epiparameter_database}{A logical that specifies whether to use the
{epiparameter} package or not.}

\item{si_mean}{A numeric that represents the mean serial interval value}

\item{si_sd}{A numeric that represents the standard deviation of the serial
interval.}

\item{si_dist}{A string used to specify the name of the serial interval
distribution. Use \code{?epiparameter::epidist} for more details about the
possible values.}
}
\value{
An object of class \if{html}{\out{<epidist>}} that contains the serial interval
distribution
}
\description{
The serial interval distribution can be either generated using the
{epiparameter} package (if the disease is covered by the package) or the
user-provided parameters.
}
\keyword{internal}
