% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmissibility.R
\name{get_transmissibility}
\alias{get_transmissibility}
\title{Estimate the disease transmissibility}
\usage{
get_transmissibility(
  data,
  disease_name,
  date_var = "date",
  count_var = "n",
  group_var = "region",
  incomplete_days = 7L,
  use_epiparameter_database = TRUE,
  si_mean = NULL,
  si_sd = NULL,
  si_dist = NULL
)
}
\arguments{
\item{data}{The input data frame or linelist}

\item{disease_name}{A string that represents the disease name}

\item{use_epiparameter_database}{A Boolean used to specify whether to extract
the serial interval distribution from the {epiparameter} package or no.
Default is \code{TRUE}.}

\item{si_mean}{A numeric representing the mean of the distribution for serial
interval. This is only used when \code{use_epiparameter_database = FALSE}.}

\item{si_sd}{A numeric representing the standard deviation of the
distribution for serial interval. This is only used when
\code{use_epiparameter_database = FALSE}.}

\item{si_dist}{A string with the name of the probability distribution for
the serial interval. This is only used when
\code{use_epiparameter_database = FALSE}.}

\item{interval}{The size of the time interval used for computing the
incidence}

\item{group_by}{A string with the name of the variable by which the data will
be grouped}

\item{method}{A string with the name of the package to be used for the
calculation of the reproduction ratio. Default is \code{i2extras}.}
}
\value{
a list with both descriptive statistics results, the growth rate, and
the reproduction number.
}
\description{
Here the transmissibility is expressed as the reproduction number i.e the
number of people infected, on average, by a single infected person.
}
