% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmissibility.R
\name{get_transmissibility_params}
\alias{get_transmissibility_params}
\title{Extract the transmissibility parameters from the input arguments}
\usage{
get_transmissibility_params(parameters, transmissibility_params)
}
\arguments{
\item{parameters}{a list of parameters to be populated with the severity
parameters}

\item{transmissibility_params}{a list with the user-specified arguments to be
used for the CFR calculation}
}
\value{
the input list of parameters with extra arguments from the
'transmissibility_params' object
}
\description{
Extract the transmissibility parameters from the input arguments
}
\keyword{internal}
