% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_templates.R
\name{list_templates}
\alias{list_templates}
\title{List available rmarkdown templates}
\usage{
list_templates()
}
\value{
A character vector containing the list of rmarkdown included in the
package
}
\description{
List available rmarkdown templates
}
\examples{
list_templates()

}
