% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_template.R
\name{make_template}
\alias{make_template}
\title{Title}
\usage{
make_template(
  data,
  disease = "COVID",
  transmissibility = TRUE,
  severity = TRUE
)
}
\arguments{
\item{severity}{}
}
\description{
Title
}
