% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmissibility.R
\name{plot_cases_per_group}
\alias{plot_cases_per_group}
\title{Plot daily incidence data}
\usage{
plot_cases_per_group(linelist_data)
}
\arguments{
\item{linelist_data}{A linelist with the raw count of the cases data}
}
\value{
An object of class \if{html}{\out{<ggplot>}}
}
\description{
Plot daily incidence data
}
\keyword{internal}
