% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmissibility.R
\name{plot_Rt}
\alias{plot_Rt}
\title{Plot timely reproduction number}
\usage{
plot_Rt(plot_data, group_var = NULL)
}
\arguments{
\item{plot_data}{A data frame with either the global or per group
reproduction number.}

\item{group_var}{A string with the name of the column used to group the data.
This is set to \code{NULL} when dealing with global reproduction number.}
}
\value{
An object of class \if{html}{\out{<ggplot>}} that contains the plot data.
}
\description{
Plot timely reproduction number
}
\keyword{internal}
