% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmissibility.R
\name{plot_si_distribution}
\alias{plot_si_distribution}
\title{Plot the serial interval distribution}
\usage{
plot_si_distribution(si, si_dist, si_mean, si_sd)
}
\arguments{
\item{si}{An object of class \if{html}{\out{<epidist>}} with the serial interval data}

\item{si_dist}{A string with the serial interval distribution name}

\item{si_mean}{A numeric with the mean serial interval}

\item{si_sd}{A numeric with the standard deviation of the serial interval}
}
\value{
An object of class \if{html}{\out{<ggplot>}}
}
\description{
Plot the serial interval distribution
}
\keyword{internal}
