% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmissibility.R
\name{plot_weekly_incidence}
\alias{plot_weekly_incidence}
\title{Plot the weekly incidence data}
\usage{
plot_weekly_incidence(converted_data)
}
\arguments{
\item{converted_data}{An object of class \if{html}{\out{<incidence2>}} with the weekly
incidence data.}
}
\value{
An object of class \if{html}{\out{<ggplot>}}
}
\description{
Plot the weekly incidence data
}
\keyword{internal}
