% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{prepare_cfr_data}
\alias{prepare_cfr_data}
\title{Get the input data for {cfr} functions}
\usage{
prepare_cfr_data(
  data,
  date_variable_name = "Onset_week",
  cases_status = "Status",
  death_outcome = "dead",
  diagnosis_status = "Type",
  diagnosis_outcome = "confirmed"
)
}
\arguments{
\item{data}{the input data}

\item{date_variable_name}{the name of the column with the dates when cases
were registered}

\item{cases_status}{the name of the column with the information about whether
a case was dead or recovered}

\item{death_outcome}{a character with the value, in the cases_status column,
that is used to specify whether the case was dead}

\item{diagnosis_status}{the name of the column with the cases diagnosis
outcome.}

\item{diagnosis_outcome}{the value, in the column with the cases diagnosis
outcome, used to represent the confirmed cases.}
}
\value{
a \code{list} of 2 elements of type data frame. These are the aggregated
data to calculate CFR across all cases and the one for CFR among the
confirmed cases only. Note that the later will be \code{NULL} if the
\code{diagnosis_status} and \code{diagnosis_outcome} are not provided.
}
\description{
Get the input data for {cfr} functions
}
\examples{
data = read.csv(system.file("extdata", "Marburg_EqGuinea_linelist.csv",
                   package = "episoap"))
cfr_data <- prepare_cfr_data(
  data               = data,
  date_variable_name = "Onset_week",
  cases_status       = "Status",
  death_outcome      = "dead",
  diagnosis_status   = "Type",
  diagnosis_outcome  = "confirmed"
)
}
\keyword{internal}
