% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{print_cfr}
\alias{print_cfr}
\title{Print the output from the CFR calculation}
\usage{
print_cfr(out)
}
\arguments{
\item{out}{the output from the CFR calculation}
}
\value{
displays the result from CFR calculation in a HTML file
}
\description{
Print the output from the CFR calculation
}
\keyword{internal}
