% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{reformat_cfr}
\alias{reformat_cfr}
\title{Print the CFR result}
\usage{
reformat_cfr(cfr)
}
\arguments{
\item{cfr}{the CFR values obtained from the \code{get_severity()} function}
}
\value{
displays the CFR result as a table
}
\description{
Print the CFR result
}
\examples{
cfr <- get_severity(
  disease_name = "Marburg Virus Disease",
  data         = readRDS(system.file("extdata",
                                      "Marburg_EqGuinea_incidence.RDS",
                                      package = "episoap")),
  account_for_delay = FALSE,
  epidist           = NULL
)
print(cfr)

}
