% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Create report from template}
\usage{
run_pipeline(
  report = list_templates(),
  out_dir = "episoap_report",
  preview = TRUE,
  ...
)
}
\arguments{
\item{report}{A character vector of report template(s) to render (default to all
templates)}

\item{out_dir}{A character vector with the output directory (default to an
\code{episoap_report} in the current directory}

\item{preview}{A logical (default \code{TRUE}) indicating whether the rendered
report should be opened in the default browser}
}
\value{
(invisibly) a character vector of paths to the rendered reports
}
\description{
Create report from template
}
\examples{
# Download data we need for this example
wd <- file.path(tempdir(), "episoap_report")
dir.create(wd)
download.file(
  "https://github.com/epiverse-trace/episoap/blob/main/inst/rmarkdown/templates/transmissibility/skeleton/data/covid_linelist_england.rds?raw=true",
  file.path(wd, "covid_linelist_england.rds")
)
# Running the pipeline with custom data saved on your computer
run_pipeline(
  report = "transmissibility",
  out_dir = wd,
  data_file = "covid_linelist_england.rds"
)
}
