% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{run_severity}
\alias{run_severity}
\title{Run severity pipeline}
\usage{
run_severity(disease_name, data, account_for_delay = TRUE, epidist = NULL, ...)
}
\description{
Run severity pipeline
}
\examples{
estimate cfr using incidence data
run_severity(
  disease_name      = "Marburg Virus Disease",
  data              = readRDS(system.file("extdata",
                                          "Marburg_EqGuinea_incidence.RDS",
                                          package = "episoap")),
  account_for_delay = FALSE,
  epidist           = NULL
)

estimate cfr using linelist data - the delay distribution parameters are
taken from the user-supplied arguments
run_severity(
  disease_name       = "Marburg Virus Disease",
  data               = read.csv(system.file("extdata",
                                            "Marburg_EqGuinea_linelist.csv",
                                            package = "episoap")),
  account_for_delay  = TRUE,
  epidist            = NULL,
  date_variable_name = "Onset_week",
  cases_status       = "Status",
  death_outcome      = "dead",
  diagnosis_status   = "Type",
  diagnosis_outcome  = "confirmed",
  distribution       = "gamma",
  type               = "range",
  values             = c(8, 2, 16),
  interval           = 1
)

run_severity(
  disease_name       = "Marburg Virus Disease",
  data               = read.csv(system.file("extdata",
                                            "Marburg_EqGuinea_linelist.csv",
                                            package = "episoap")),
  account_for_delay  = TRUE,
  epidist            = NULL,
  date_variable_name = "Onset_week",
  cases_status       = "Status",
  death_outcome      = "dead",
  diagnosis_status   = "Type",
  diagnosis_outcome  = "confirmed",
  distribution       = "lnorm",
  meanlog            = 2.5,
  sdlog              = 1.2
)

estimate cfr using count data
run_severity(
  disease_name       = "Marburg Virus Disease",
  data               = NULL,
  account_for_delay  = FALSE,
  epidist            = NULL,
  total_cases        = 70,
  total_deaths       = 25,
  death_in_confirmed = 12
)
}
