% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_episoap.R
\name{theme_episoap}
\alias{theme_episoap}
\title{A custom \pkg{ggplot2} theme for \pkg{episoap} reports}
\usage{
theme_episoap()
}
\value{
A \pkg{ggplot2} \link{theme} object
}
\description{
A custom \pkg{ggplot2} theme for \pkg{episoap} reports
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  labs(title = "Fuel economy declines as weight increases") +
  theme_episoap()

}
