% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_soap.R
\name{theme_soap}
\alias{theme_soap}
\title{A custom \pkg{ggplot2} theme for \pkg{soap} reports}
\usage{
theme_soap()
}
\value{
A \pkg{ggplot2} \link{theme} object
}
\description{
A custom \pkg{ggplot2} theme for \pkg{soap} reports
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  labs(title = "Fuel economy declines as weight increases") +
  theme_soap()

}
