% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmissibility.R
\name{wrap_res}
\alias{wrap_res}
\title{Process output from \code{EpiEstim::estimate_R()} and output a tibble with dates,
mean R, and associated confidence intervals.}
\usage{
wrap_res(x, incid = NULL)
}
\arguments{
\item{x}{An \code{estimate_R} object}

\item{incid}{(Optional) The incidence2 object on which R values have been
estimated.}
}
\value{
An object of class <R_estimate> and \if{html}{\out{<tibble>}}.
}
\description{
Process output from \code{EpiEstim::estimate_R()} and output a tibble with dates,
mean R, and associated confidence intervals.
}
\keyword{internal}
