% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmissibility.R
\name{wrap_si}
\alias{wrap_si}
\title{Process a distrcrete object and output \code{si_discr} argument for
EpiEstim::make_config}
\usage{
wrap_si(si)
}
\arguments{
\item{si}{An object of class \if{html}{\out{<epidist>}} with the serial interval information}
}
\value{
A numeric vector of the serial interval qvalues
}
\description{
Process a distrcrete object and output \code{si_discr} argument for
EpiEstim::make_config
}
\keyword{internal}
