% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{calculate_cfr}
\alias{calculate_cfr}
\title{Calculate CFR from incidence or linelist data}
\usage{
calculate_cfr(
  data,
  epidist = NULL,
  account_for_delay = FALSE,
  type,
  values,
  distribution,
  interval,
  shape,
  scale,
  meanlog,
  sdlog
)
}
\arguments{
\item{data}{the input data}

\item{epidist}{an \code{epidist} object that contains the distribution parameters}
}
\value{
a \code{list} of the following 2 elements:
\enumerate{
\item cfr: the CFR among the total cases
\item cfr_in_confirmed_cases: the CFR among the confirmed cases only
}
}
\description{
Calculate CFR from incidence or linelist data
}
\examples{
cfr_data <- prepare_cfr_data(
  data               = data,
  date_variable_name = "Onset_week",
  cases_status       = "Status",
  death_outcome      = "dead",
  diagnosis_status   = "Type",
  diagnosis_outcome  = "confirmed"
)
cfr <- calculate_cfr_from_incidence(
  data              = cfr_data,
  epidist           = NULL
)
}
\keyword{internal}
