% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{calculate_cfr_from_counts}
\alias{calculate_cfr_from_counts}
\title{Calculate CFR from count data}
\usage{
calculate_cfr_from_counts(total_cases, total_deaths, death_in_confirmed)
}
\arguments{
\item{total_cases}{a \code{numeric} that represent the total number of cases in
the data}

\item{total_deaths}{a \code{numeric} that represents the total number of deaths in
the data}

\item{death_in_confirmed}{a \code{numeric} that represents the number of death in
among the confirmed cases.}
}
\value{
an object of type \code{list} with 2 elements of type \verb{data frame}.
}
\description{
Calculate CFR from count data
}
\keyword{internal}
