% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{calculate_cfr_from_incidence}
\alias{calculate_cfr_from_incidence}
\title{Estimate disease severity from incidence object}
\usage{
calculate_cfr_from_incidence(data, epidist, epidist_params)
}
\arguments{
\item{data}{the input data}

\item{epidist}{an object of class \code{epidist} with the delay distribution
parameters}
}
\value{
an object of type \code{list} with 2 data frames that contains the
CFR values in all cases and in confirmed cases only.
}
\description{
Estimate disease severity from incidence object
}
\keyword{internal}
