% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{calculate_cfr_from_linelist}
\alias{calculate_cfr_from_linelist}
\title{Estimate disease severity from linelist object}
\usage{
calculate_cfr_from_linelist(data, epidist = NULL, epidist_params = NULL)
}
\arguments{
\item{data}{the input data}

\item{epidist}{an object of class \code{epidist} with the delay distribution
parameters}

\item{epidist_params}{A list with the parameters to be used to get the delay
distribution. Possible values are:
\describe{
type, values, distribution, interval, shape, scale, meanlog, sdlog
}}
}
\value{
an object of type \code{list} with 1 or 2 data frames that contains the
CFR values in all cases and in confirmed cases only.
}
\description{
Estimate disease severity from linelist object
}
\keyword{internal}
