% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{calculate_overall_cfr}
\alias{calculate_overall_cfr}
\title{Calculate CFR with uncertainty}
\usage{
calculate_overall_cfr(
  data,
  infection_type = "direct_contact",
  death_status = "Status",
  case_status = NULL
)
}
\arguments{
\item{data}{a data frame or a linelist object}

\item{infection_type}{the type of the infection}

\item{death_status}{the name of the column with the case's death status}

\item{case_status}{}

\item{onset_date}{the name of the column with the onset date values}

\item{date_scale}{the time scale of the onset date values}
}
\description{
Calculate CFR with uncertainty
}
