% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{create_config}
\alias{create_config}
\title{Create and Open a Configuration YAML File Template}
\usage{
create_config(path = file.path(tempdir(), "config.yaml"))
}
\arguments{
\item{path}{A character with the file path where the configuration file
should be created/opened. When not provided, default to a file named
\code{config.yaml} in the system's temporary directory \code{tempdir()}.}
}
\value{
Invisibly returns the \code{path} to the created or existing
configuration file.
}
\description{
This function generates a configuration file in a YAML format with a
structured template for epidemiological severity parameters. If the file does
not exist at the specified path, it creates one with placeholder values. The
file is then opened in the system's default application for YAML/text files.
}
\note{
\itemize{
\item Users need to replace \code{NA} with appropriate values before using the configuration.
\item The function does not validate YAML content.
}
}
\examples{
# Create/open config in temp directory
create_config()

# Create/open in working directory
create_config("my_config.yaml")

}
