% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{get_onset_to_death_distro}
\alias{get_onset_to_death_distro}
\title{Create the delay distribution object}
\usage{
get_onset_to_death_distro(data, disease = "Marburg Virus Disease", ...)
}
\arguments{
\item{data}{the input data frame or linelist object}

\item{disease}{the name of the disease of interest}

\item{...}{the other extra arguments. Possible values are:
\enumerate{
\item type: a \code{character} specifying whether summary statistics are based
around percentiles (default) or range. Type \code{?epiparemeter::extract_param} # nolint: line_length_linter
for more details.
\item values: a numeric \code{vector}. Type \code{?epiparemeter::extract_param} for
more details
\item distribution: a \code{character} string that specifies the distribution to
use. Type \code{?epiparemeter::extract_param} for more details.
\item shape: the shape of the specified distribution (when distribution = gamma or weibull) # nolint: line_length_linter
\item scale: the scale of the specified distribution (when distribution = gamma or weibull) # nolint: line_length_linter
\item meanlog: the meanlog of the specified distribution (when distribution = lnorm) # nolint: line_length_linter
\item sdlog: the sdlog of the specified distribution (when distribution = lnorm) # nolint: line_length_linter
\item interval: the interval to discretise the interval onto. Default is 1
for daily case and death data.
}}
}
\value{
an object of type \code{epidist} with the epidemiological parameters of
the disease of interest.
}
\description{
Create the delay distribution object
}
\examples{
onset_death <- get_onset_to_death_distro(
  data         = data,
  disease      = "Marburg Virus Disease",
  type         = "range",
  values       = c(8, 2, 16),
  distribution = "gamma"
)
}
\keyword{internal}
