% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{get_sequential_dates}
\alias{get_sequential_dates}
\title{Convert non-daily incidence data into daily incidence data}
\usage{
get_sequential_dates(data)
}
\arguments{
\item{data}{A \code{<data.frame>} with at least two columns named as \code{cases}
and the \code{date}}
}
\value{
A \code{<data.frame>} with three columns named as 'date', 'cases',
'deaths'. This is the format required by the \code{cfr::cfr_static()}
function.
}
\description{
Convert non-daily incidence data into daily incidence data
}
\examples{
weekly_incidence <- readRDS(
  system.file("extdata", "weekly_incidence.RDS", package = "episoap")
)
daily_incidence <- get_sequential_dates(weekly_incidence)

}
