% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_config.R
\name{load_config}
\alias{load_config}
\title{Load and Validate Configuration Parameters From Yaml}
\usage{
load_config(path = file.path(tempdir(), "config.yaml"))
}
\arguments{
\item{path}{Path to config file (default: tempdir()/config.yaml)}
}
\value{
List of validated parameters from the yaml
}
\description{
Load and Validate Configuration Parameters From Yaml
}
