% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity.R
\name{print_cfr}
\alias{print_cfr}
\title{Print the CFR result}
\usage{
print_cfr(cfr)
}
\arguments{
\item{cfr}{the CFR values obtained from the \code{get_severity()} function}
}
\value{
displays the CFR result as a table
}
\description{
Print the CFR result
}
\examples{
cfr <- get_severity(
  data              = read.csv(system.file("extdata",
                                           "Marburg_EqGuinea_linelist.csv",
                                            package = "episoap")),
  infection_type    = "direct_contact",
  cases_status      = "Status",
  outcomes          = c("dead", "recovered"),
  diagnosis_status  = "Type",
  diagnosis_outcome = "confirmed"
)
print(cfr)

}
