% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Execute the user-defined tasks}
\usage{
run_pipeline(
  disease_name,
  data,
  to_incidence_params = NULL,
  severity_params = NULL,
  transmissibility_params = NULL
)
}
\arguments{
\item{disease_name}{A string with the disease name}

\item{data}{A data frame or linelist or incidence object}

\item{to_incidence_params}{A list of parameters needed to convert the input
data into an incidence object. This arguments is only needed when the user
wants to include severity into the run and the data is a linelist or data
frame.}

\item{severity_params}{A list with the parameters required calculate the
disease severity.}

\item{transmissibility_params}{A list with the parameters needed to estimate
the disease transmissibility}
}
\value{
Generates a HTML file with the result from each element of the
pipeline
}
\description{
Execute the user-defined tasks
}
\examples{
data <- outbreaks::ebola_kikwit_1995 |>
  dplyr::rename("cases" = "onset", "deaths" = "death")
disease_name <- "ebola"

# define the transmissibility arguments
define the arguments for transmissibility
transmissibility_params <- list(
  date_var = "date",
  group_var  = NULL,
  count_var = "cases",
  incomplete_days = 7,
  use_epiparameter_database = TRUE,
  si_mean       = NULL,
  si_sd         = NULL,
  si_dist       = NULL
)

# define the severity arguments
to_incidence_params <- NULL
severity_params <- list(
  account_for_delay  = FALSE,
  epidist            = NULL,
  type               = "range",
  values             = c(8, 2, 16),
  distribution       = "gamma",
  interval           = 1,
  meanlog            = NULL,
  sdlog              = NULL,
  total_cases        = NULL,
  total_deaths       = NULL,
  death_in_confirmed = NULL
)

# run the pipeline
run_pipeline(
  disease_name = disease_name,
  data = data,
  to_incidence_params = to_incidence_params,
  severity_params = severity_params,
  transmissibility_params = transmissibility_params
)
}
